// syncro/apps/web/src/lib/api/generated/model/installationProjection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */
import type { InstallationProjectionReason } from './installationProjectionReason';

export interface InstallationProjection {
  installationId?: string;
  sparepartId?: string;
  functionName?: string;
  available?: boolean;
  reason?: InstallationProjectionReason;
  remainingCounters?: number;
  projectedDepletionAt?: string;
  leadTimeHours?: number;
  consumptionDuringLeadTime?: number;
}