// syncro/apps/web/src/lib/api/generated/model/machineSparepartProjectionsView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */
import type { InstallationProjection } from './installationProjection';
import type { MachineSparepartProjectionsViewCalculationBasis } from './machineSparepartProjectionsViewCalculationBasis';
import type { MachineSparepartProjectionsViewInsufficientReason } from './machineSparepartProjectionsViewInsufficientReason';

export interface MachineSparepartProjectionsView {
  machineId?: string;
  rateAvailable?: boolean;
  calculationBasis?: MachineSparepartProjectionsViewCalculationBasis;
  windowStartAt?: string;
  windowEndAt?: string;
  firstSampleAt?: string;
  lastSampleAt?: string;
  insufficientReason?: MachineSparepartProjectionsViewInsufficientReason;
  ratePerOperatingHour?: number;
  shiftSource?: string;
  dailyOperatingHours?: number;
  projections?: InstallationProjection[];
}